✨ Give a Kinetic one rectangle, and judge everything against it
The map was the only thing an entity could run into. Everything else a cart
does with a rectangle -- a bullet against an aircraft, a player against a
pickup, either against the edge of the screen -- was hand-rolled arithmetic
over four coordinates, once per cart.

`Bounds` is that rectangle and `bounds` is where an entity says which one is
its. `Collider` goes with it: it held a size while `resolve` was handed the
position and the velocity, which left it half a shape and half an actor, and
the size it held was a second rectangle to keep in step with the first. What
survives is the part a cart actually chooses -- `solid`, the sprite flags that
mean wall to this entity -- with the box coming from `bounds` like everything
else. `MapCollider` pairs the two behind the module wall for as long as one
update takes to resolve, and the contacts move out to a module of their own
since the edge of the world reports them too.

Note the rectangle is measured from the drawn position rather than the exact
sub-pixel one the map resolution keeps to: what the player sees overlapping
has to overlap, while a wall shifting under an entity by a pixel would judder.
The two are reconciled where the collider is built, so an entity whose
rectangle sits where its body draws meets a wall at exactly the position it
always did.

The platformer in this tree is the one cart that had two boxes -- a forgiving
8x7 for the badie and the whole 8x8 for the walls -- and it keeps the one the
walls stop. Its badie check is still the hand-rolled arithmetic here and reads
the same constant, so for this one commit a stomp registers a row earlier than
it did; the next commit takes that back when the check becomes
`Bounds::overlaps`, leaving the vertical window exactly where it started.

+++ examples/platformer/src/lib.rs
@@ -110,8 +110,8 @@ impl Platformer {
 
         // Check for collision between our hero and the badie, and decide who dies if there is one.
         if let Some(badie) = &mut self.badie {
-            if self.hero.draw_x() + HERO_WIDTH >= badie.draw_x()
-                && self.hero.draw_x() < badie.draw_x() + BADIE_WIDTH
+            if self.hero.draw_x() + HERO_WIDTH as i16 >= badie.draw_x()
+                && self.hero.draw_x() < badie.draw_x() + BADIE_WIDTH as i16
             {
                 if self.hero.draw_y() == badie.draw_y() {
                     // Hero ramming into badie horizontally is a suicide.
@@ -119,8 +119,8 @@ impl Platformer {
                     self.game_over(ctx);
 
                     return;
-                } else if self.hero.draw_y() + HERO_HEIGHT >= badie.draw_y()
-                    && self.hero.draw_y() < badie.draw_y() + BADIE_HEIGHT
+                } else if self.hero.draw_y() + HERO_HEIGHT as i16 >= badie.draw_y()
+                    && self.hero.draw_y() < badie.draw_y() + BADIE_HEIGHT as i16
                 {
                     // Hero hitting the badie from the top, kills the badie and gives hero a boost.
                     self.badie = None;
